+++ fastvideo/models/dits/matrix_game/action_module.py
@@ -73,8 +73,6 @@ def _apply_rotary_emb_qk(
 class ActionModule(nn.Module):
     """
     action module from https://arxiv.org/pdf/2501.08325
-    鼠标控制信号的输入是一个 L*D 的向量
-    键盘同样
     """
 
     def __init__(
