+++ fastvideo/pipelines/basic/matrixgame/__init__.py
@@ -1,0 +1,7 @@
+# SPDX-License-Identifier: Apache-2.0
+from fastvideo.pipelines.basic.matrixgame.matrixgame_i2v_pipeline import (
+    MatrixGamePipeline)
+from fastvideo.pipelines.basic.matrixgame.matrixgame_causal_dmd_pipeline import (
+    MatrixGameCausalDMDPipeline)
+
+__all__ = ["MatrixGamePipeline", "MatrixGameCausalDMDPipeline"]
