+++ fastvideo/pipelines/basic/matrixgame/matrixgame_i2v_pipeline.py
@@ -1,32 +1,27 @@
 # SPDX-License-Identifier: Apache-2.0
-"""
-Wan video diffusion pipeline implementation.
-
-This module contains an implementation of the Wan video diffusion pipeline
-using the modular pipeline architecture.
-"""
+"""Matrix-Game I2V pipeline implementation."""
 
 from fastvideo.fastvideo_args import FastVideoArgs
 from fastvideo.logger import init_logger
 from fastvideo.pipelines.composed_pipeline_base import ComposedPipelineBase
 from fastvideo.pipelines.lora_pipeline import LoRAPipeline
 
-# isort: off
 from fastvideo.pipelines.stages import (
-    ImageEncodingStage, ConditioningStage, DecodingStage, DenoisingStage,
-    ImageVAEEncodingStage, InputValidationStage, LatentPreparationStage,
-    TextEncodingStage, TimestepPreparationStage)
-# isort: on
+    MatrixGameImageEncodingStage, ConditioningStage,
+    DecodingStage, DenoisingStage,
+    InputValidationStage, LatentPreparationStage, TextEncodingStage,
+    TimestepPreparationStage)
+from fastvideo.pipelines.stages.image_encoding import (
+    MatrixGameImageVAEEncodingStage)
 from fastvideo.models.schedulers.scheduling_flow_unipc_multistep import (
     FlowUniPCMultistepScheduler)
 
 logger = init_logger(__name__)
 
 
-class WanImageToVideoPipeline(LoRAPipeline, ComposedPipelineBase):
-
+class MatrixGamePipeline(LoRAPipeline, ComposedPipelineBase):
     _required_config_modules = [
-        "text_encoder", "tokenizer", "vae", "transformer", "scheduler", \
+        "vae", "transformer", "scheduler",
         "image_encoder", "image_processor"
     ]
 
@@ -35,22 +30,22 @@ class WanImageToVideoPipeline(LoRAPipeline, ComposedPipelineBase):
             shift=fastvideo_args.pipeline_config.flow_shift)
 
     def create_pipeline_stages(self, fastvideo_args: FastVideoArgs):
-        """Set up pipeline stages with proper dependency injection."""
-
         self.add_stage(stage_name="input_validation_stage",
                        stage=InputValidationStage())
 
-        self.add_stage(stage_name="prompt_encoding_stage",
-                       stage=TextEncodingStage(
-                           text_encoders=[self.get_module("text_encoder")],
-                           tokenizers=[self.get_module("tokenizer")],
-                       ))
+        if (self.get_module("text_encoder", None) is not None
+                and self.get_module("tokenizer", None) is not None):
+            self.add_stage(stage_name="prompt_encoding_stage",
+                           stage=TextEncodingStage(
+                               text_encoders=[self.get_module("text_encoder")],
+                               tokenizers=[self.get_module("tokenizer")],
+                           ))
 
-        if (self.get_module("image_encoder") is not None
-                and self.get_module("image_processor") is not None):
+        if (self.get_module("image_encoder", None) is not None
+                and self.get_module("image_processor", None) is not None):
             self.add_stage(
                 stage_name="image_encoding_stage",
-                stage=ImageEncodingStage(
+                stage=MatrixGameImageEncodingStage(
                     image_encoder=self.get_module("image_encoder"),
                     image_processor=self.get_module("image_processor"),
                 ))
@@ -68,16 +63,17 @@ class WanImageToVideoPipeline(LoRAPipeline, ComposedPipelineBase):
                            transformer=self.get_module("transformer")))
 
         self.add_stage(stage_name="image_latent_preparation_stage",
-                       stage=ImageVAEEncodingStage(vae=self.get_module("vae")))
+                       stage=MatrixGameImageVAEEncodingStage(
+                           vae=self.get_module("vae")))
 
         self.add_stage(stage_name="denoising_stage",
                        stage=DenoisingStage(
                            transformer=self.get_module("transformer"),
-                           transformer_2=self.get_module("transformer_2"),
+                           transformer_2=self.get_module("transformer_2", None),
                            scheduler=self.get_module("scheduler")))
 
         self.add_stage(stage_name="decoding_stage",
                        stage=DecodingStage(vae=self.get_module("vae")))
 
 
-EntryClass = WanImageToVideoPipeline
+EntryClass = [MatrixGamePipeline]
