+++ fastvideo/pipelines/pipeline_registry.py
@@ -26,8 +26,8 @@ _PIPELINE_NAME_TO_ARCHITECTURE_NAME: dict[str, str] = {
     "StepVideoPipeline": "stepvideo",
     "HunyuanVideoPipeline": "hunyuan",
     "Cosmos2VideoToWorldPipeline": "cosmos",
-    "MatrixGamePipeline": "wan",
-    "MatrixCausalGameDMDPipeline": "wan",
+    "MatrixGamePipeline": "matrixgame",
+    "MatrixGameCausalDMDPipeline": "matrixgame",
 }
 
 _PREPROCESS_WORKLOAD_TYPE_TO_PIPELINE_NAME: dict[WorkloadType, str] = {
